fix building errors

+++ thinlibcxx/cstdint.hpp
@@ -9,5 +9,5 @@ using int16_t = short int;
 using uint8_t = unsigned char;
 using int8_t = char;
 
-using size_t = unsigned long long int;
+using size_t = unsigned long int;
 using size_p = unsigned long long int;
